feat(profile): move approver profile control to header
