+++ dashboard/src/app.css
@@ -47,12 +47,12 @@
 }
 
 @layer components {
-  .sidebar-button-toggle {
+  .control-button {
     @apply p-1 rounded text-gray-500;
     @apply transition-colors duration-300;
   }
 
-  .sidebar-button-toggle:hover {
+  .control-button:hover {
     @apply bg-black bg-opacity-5;
   }
 }
