+++ dashboard/src/lib/components/navs/sidebar.svelte
@@ -1,35 +1,11 @@
 <script lang="ts">
   import type { ComponentType } from "svelte"
-  import type { Approver } from "$lib/types"
-  import { approver } from "$lib/store"
   import { fade, slide } from "svelte/transition"
   import Wordmark from "$lib/components/utils/wordmark.svelte"
-  import BasicModal from "$lib/components/modals/basic.svelte"
-  import InputField from "$lib/components/inputs/field.svelte"
-  import BasicButton from "$lib/components/buttons/basic.svelte"
   import { SidePanelCloseFilled, Home, LogoGithub } from "carbon-icons-svelte"
 
   export let open: boolean = true
   export let close: () => void
-
-  let showApproverModal = !$approver
-  let approverName = ""
-  let approverEmail = ""
-
-  function saveProfile(name: string, email: string) {
-    if (!name || !email) {
-      return
-    }
-
-    let profile: Approver = {
-      name: name,
-      email: email
-    }
-
-    approver.set(profile)
-    window.localStorage.setItem("approver", JSON.stringify($approver))
-    showApproverModal = false
-  }
 
   type Navigation = {
     icon: ComponentType
@@ -61,7 +37,7 @@
   <nav class="sidebar" transition:slide={{ axis: "x", duration: 250 }}>
     <div class="control-section space-x-4">
       <Wordmark size="sm" />
-      <button class="sidebar-button-toggle" on:click={close}>
+      <button class="control-button" on:click={close}>
         <SidePanelCloseFilled size={24} />
       </button>
     </div>
@@ -79,35 +55,6 @@
     </div>
   </nav>
 {/if}
-
-<BasicModal bind:show={showApproverModal}>
-  <div class="flex flex-col space-y-6">
-    <div class="flex flex-col space-y-3">
-      <h3>Approver Profile</h3>
-      <small>
-        This information adds more context to the approval response.
-      </small>
-    </div>
-    <div class="flex flex-col space-y-3">
-      <InputField
-        id="name"
-        label="Full Name"
-        placeholder="Justin Case"
-        bind:value={approverName}
-      />
-      <InputField
-        id="email"
-        label="Email Address"
-        placeholder="justincase@example.com"
-        bind:value={approverEmail}
-      />
-    </div>
-    <BasicButton
-      text="Save Profile"
-      action={() => saveProfile(approverName, approverEmail)}
-    />
-  </div>
-</BasicModal>
 
 <style lang="postcss">
   .sidebar {
